+++ CHANGELOG.md
@@ -4,7 +4,6 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Ideas/Upcoming]
-Pending changes
 Switch Animations
 
 ## [Unreleased]
@@ -13,6 +12,7 @@ Switch Animations
 * ColorPicker
 * FontPicker
 * Support control caching in the NavigationPresenter
+* Support user control switch abort by pending changes and other reason
 
 ## [1.0.0] - 2021-06-13
 ### Added